Update
Work on Packages controller and views

+++ Panda/Panda.App/Models/Package/PackageCreateBindingModel.cs
@@ -9,10 +9,10 @@ namespace Panda.App.Models.Package
     public class PackageCreateBindingModel
     {
         public string Description { get; set; }
-
+         
         public string ShippingAddress { get; set; }
 
-        public decimal Weight { get; set; }
+        public double Weight { get; set; }
 
         public string Recipient { get; set; }
